working on BotSmart
Implemented to a stage where BotSmart can export the decision tree
(using .graphml format that the open source software yEd can open and
layout nicely, just mark all nodes and go to Layout > Tree for instance)
one level deep (meaning all options for the next move), which is of
course not enough - it'll have to be build out with recursions and also
depth-limits to avoid high computation time.
Also some improvements in Board and Field. For instance the method to
collect free fields is build into Field now and uses the SnakePath to
walk along all fields while collecting.