more work on BotSmart [#9]
Moved DecisionTree-related classes into a tree-package within the
bot-package.
The building of the tree works now, but no judgement is taken yet as to
which branches to give what weight to (highest weight will be chosen). I
was overwhelmed by how very quickly the graph grows... had to limit it
to three turns deep (on a 3x3 field!!) to oversee things... will paste a
screenshot into #9.
So some formula for predicting and restricting the computation-load will
be absolutely necessary.
Also spotted and solved a hidden error in Field that would have caused
wrong freeFieldCounts when the 0,0-field is taken.