working on the decision tree [#9]
Now it doesn't create children if a winState (= win or loose for the
smartBot who is "thinking") of a node is reached (checked through
Berni's BoardParser. In the export you can see that win-nodes are
colored blue and loose-nodes are red.
Next are the branch-weights to actually make the decision, not just look
at all possible decision :)