[#9] smartBot makes choices now!
But, they don't seem overly smart yet :) more work to do...
Sometimes it crashes when the chosenIndex (by default -1) is not
overwritten with a better one for some reason. And sometimes the "smart"
bot doesn't stop me from winning in the next round... definitely more
work on the smartness required. And everything bigger than 3x3 causes
out-of-memory currently because no dynamic workload-forecasting is
implemented yet.