working on smartBot
Outsourced all the "weight"-giving into the static method of
WeightController as that allows for better experimenting with the
parameters.
The smartBot is very smart in many situations now - but also annoyingly
stupid in other situations. Haven't found a way to consolidate immediate
win possibilites with the high sums resulting from adding up all kinds
of branches.
Maybe a bigger conceptual change has to happen. Maybe TreeBuilder should
only be created once instead of re-created with every turn of smartBot.
#9