turnIndize-array out / improving WeightController
I got rid of the strange turnIndize-array and am now computing it as we
go down the levels.
Also improved the WeightController and have settled for now on
parameters that give more weight to avoiding loosing and the experience
playing against smartBot is quite exciting now i would say :)
Having two smartBots play against each other is very amusing to watch...
and it does still expose weeknesses in the decision process. I wonder if
it's a matter of finding the best set of parameters or if a different
concept alltogether is the way to make it ultra-smart...
#9