better info in console / bonus for immediate win
The info one get's in the console is a bit richer now: added the number
of nodes and edges created, the time in milliseconds it took to
"decide", markCount within marksPerTurn is more visible and among how
many equal rootnode children one is chosen.
Also i added a "hyperbonus" (Integer.MAX_VALUE) in WeightController if
there is an immediate win possible. It's not elegant though... i might
comment it out later again :)
Also put DemoMain back into the treePackage. It's commented out - just
in case someone is interested in seeing how i used a Main from the
"side" (not firing up the whole app every time) during the smartBot
development.