QuickMode in Model [#26] / depth 6 for 3x3
Implememented QuickMode using the interval-variable that was already
reserved for that. 0 means no time limit for HumanLocal - any other
number is the time-limit in seconds. Used new class Timer as a clock.
Increased default tree-depth for 3x3 to 6 to give my smartBot the
ability to foresee a tricky play (on 3x3) that Berni discovered where
both of our smartBots were loosing:
human starting with 0,0, bot 1,1, human 2, 2, bot either 0,2 or 2,0
(randomly, had same value at depth 5) and then if human chooses 0,2 the
bot has certainly lost. Now with depth 6 bot places at 0,2 and doesn't
give human the chance for that trick :)